Added backbone template importing mechanism,
In general this mechanism works off of <link type="text/html" rel="import" id="x" href="/templates/x.html" /> tags
where type should be "type/html"
      rel must be "import"
      id is normally "x-template"
      href is "/template/x.html"

for this commit I used an example template called banner
note that the tag is <link type="text/html" rel="import" id="banner-template" href="/templates/banner.html" />

these will come in handy when we start working with backbone.

For future reference

To make a template
1. put a file into Market/app/views/templates/$NAME.html{,.erb}
2. add a method to TemplatesController with the name $NAME

To use a template
1. when initializing views for the first time, wrap them in this
templateLoader.on("loaded", function() {
  // code here
});
this is subject to change, but it gets around an async issue when loading a template compared to initializing a template.